printer_supply: use supply toner color metric
The following improvements were made to the plugin:

  - drop generic "pages" metric as it's not relevant to the printer
    supply level data coming from the SNMP address.
  - use a supply toner metric based on the color code provided by the
    data extracted from the service description.
  - improve the summary in the check levels metric and only show
    unit-based summary when the unit is not a percentage.

Note: this change will drop the legacy "pages" metric and the generic
"supply_toner" metric. It's not possible to migrate "supply_toner" to
the new `supply_toner_{color}` metric. Since we'd also need the ability
to translate the metric based on color which isn't possible.

CMK-22267

Change-Id: I4cfc6371471e8765e952b987886faa07564c366a