16575 FIX Nagios core: Fix Invalid host object directive 'service_period'
Fix a bug in the core configuration generation that caused an error when the
"Service period for services" or "Service period for hosts" rule sets were
configured with the Checkmk Raw Edition. In this case activate changes, more
specifically the computation of the core configuration failed with the
following error:

```
Error: Invalid host object directive 'service_period'".
```

The regression was caused by some refactoring (52362a3e) which moved
some patching of the global configuration to some specific place
where it was not executed in all relevant code paths anymore.

Besides fixing the bug, the change also cleans up the patching of the
configuration, which could lead to subtle bugs.

CMK-22536

Change-Id: I1285b7f448287fd035b703847ea06cb39379abd9