17729 FIX Fix misbehaving check plugins after a reboot
The check plugins are allowed to store data for the next check cycle via the `value store`.
During `omd umount` these data structures are persisted.
Due to a regression introduced in [Werk #15872](https://checkmk.com/werk/15872), the same data would be loaded after every reboot.
This meant that check plugins would receive data, which was not from the previous check cycle, but rather from the last time `omd umount` was ran.

In the 2.4 this could result in the following exception
```
JSONDecodeError (Expecting property name enclosed in double quotes: line 1 column 2 (char 1))
```

The same issue also affects piggyback data.
With this Werk, the use of outdated data  is fixed.

Note, after a reboot the `value store` is likely to be empty.
The check plugins need to correctly initialize their data on the first check cycle.

Change-Id: I4907119c504918d7a0991d440be5f814b052207c