plesk plugins: Update man page and error message
The agent plugins `plesk_domains` and `plesk_backups`
both require the Python package `MySQLdb`. But neither
the man pages nor the error message gave a user clear
instructions on how to install it.
With this change example commands are added.

Also, the error messages about the missing Python package
now ends with a newline. This way, a following section
will start in a new line so that it can be parsed.

SUP-23018

Change-Id: I4cc8abafaaf204ee388d9ae069adeb0cee489e0e