17860 FIX Azure Quick Setup: Handle proxy connection errors
On the first stage of the Azure Quick Setup a connection test is done. Prior to this Werk the UI would not show an error in case the connection to the proxy was unsuccessful.

This behavior is fixed. An error message is shown whenever the Azure connection test runs into a proxy error.

This also holds for the Azure special agent when run on the command line with option "--connection-test". The special agent fails with return code 2 and shows the same error message that is rendered to the UI for the Quick Setup.

Change-Id: Id2e5cbd37d43b498116a24372400a0cc0b8c1742