15359 FIX move dell_powerconnect to exchange.checkmk.com
You are affected by this change if you use one of the following
dell_powerconnect services:

* dell_powerconnect_cpu
* dell_powerconnect_fans
* dell_powerconnect_psu
* dell_powerconnect_temp

They were removed from Checkmk because of conflicting licenses.

The code was packaged as a MKP and is available from the exchange:

https://exchange.checkmk.com/p/dell-powerconnect

Install this package if you want to continue monitoring dell powerconnect
devices.

Change-Id: I3bc55b67258ba19e0b1f6f2c2a55ca477a53100a