Attempt at stabilizing `test_mkbackup_locking` integration test
This test fails every now and then in various ways. Example:
```
psutil.NoSuchProcess: process PID not found (pid=27298)
```
The failures are usually related to the file locking process started by the
test. We try to stabilize the test by re-implementing this part and making the
interactions with the file locking process more explicit.

Change-Id: I3451b9665425c7906337df813b3c483eea31ba56