docker update test: Remove global constant and skipif
The docker update test was using a global constant for the base
release and a skipif based on git logic to skip the test in case
no tag was found for that base release. This logic is unreliable
since it depends on having a full checkout, even though checking
out anything else apart from the current branch is unnecessary.

Therefore, this change removes the skipif and moves the global
constant to the test function.

Change-Id: Ib131f1e94afab00e29bcb2319fca944745559d88