17533 Add disabled schedule status to MS SQL Job service
Previously, if the schedule for a job was disabled,
the status of the service for such an enabled job
was determined by the status of the disabled job.

Since this release, MS SQL Server Job's rule now also
have a status for cases with disabled schedules,
which is set to OK by default.

Change-Id: I7f60c482d57acec1792c6fbfa0ade36dd1ef0e6e