17731 FIX omd update: Notify about unacknowledged Werks only on central site
Checkmk keeps track of unacknowledged Werks per site.
During an update to a new major version these Werks are automatically acknowledged and the list of unacknowledged Werks is replaced with the ones of the new version.
Due to this behaviour, `omd update` warns users before the update if there are any unacknowledged Werks.

With this Werk, the update warning only applies on the central site. Previously, it only applied on remote sites.

Change-Id: I8ebb73faf6fdbc10c5f5074a7db849300886984c