17837 FIX printer supply perfometer
When a black printer supply is detected, the perfometer background is
set to black which is illegible. The background color is now changed to
a dark gray. In addition, we're now setting the default background to
brown, when a supply toner is not found. This is consistent with future
versions of Checkmk.

CMK-14950

Change-Id: Ic50d3ba506ef5ad5cd06897d6d1b9b46eb95e867