Avoid `UnboundLocalError` when rendering icon
We encountered this in the context of SUP-23016. The corresponding werk made
both the original error and the error fixed in this change go away in the setup
described in the ticket. Since we don't know if and under what conditions the
error fixed here would occur on its own, we don't write a werk.

Change-Id: I9ebe82d617fdff62414a557d5b291e756f710d16