17818 FIX UpdateAction: only remove user profiles that fail validation
In a previous werk (#17817) we introduced an update action that would
remove any user profiles that weren't also current Checkmk users, in
an attempt to clean up user IDs that weren't valid. However, this change
didn't check the validation of these user IDs which was incorrect.

With this werk we now check first if the user profile user ID is
not valid, and only then delete it.

Change-Id: Id10056e0abed7e0efb8536402f00d582de2c086c