17512 REST API: change endpoint method for list service endpoints
The REST API endpoints "Show all monitored services" and
"Show the monitored services of a host" were previously accessible via the GET
method. This caused a limitation of the maximum size of query parameters.

A common practice is bypass this limitation is to use a POST request. Other
werks like 17003 have already implemented this change for some other endpoints.

This werk changes the two aforementioned endpoints to POST.
The GET versions will continue to work until Checkmk version 2.5.

Change-Id: I63799b187a899fab496d97dcc8cfcdf1b0e668fd
JIRA-Ref: CMK-22009