14845 Stop shipping "doc/treasures" with omd
We keep `doc/treasures` in the git repository, as code in `doc/treasures` has been useful for our community repeatedly. Scripts, plugins, or other tools are collected here because they are useful, but they are not part of the main product and therefor no guarantee regarding functionality nor support can be given.

CMK-22580

Change-Id: I1f49b0a8946b7fbfd3cbaca07812b97009c71190