Deploy automation helper with uvicorn instead of gunicorn
When gunicorn forks off its worker processes, all inotify instances created by
the watcher thread are duplicated. This should not be a problem, but it is still
cleaner to avoid this by switching to uvicorn, which spawns workers instead.
Also, this reduces the number of dependencies, since we anyway depend on uvicorn
for the worker class when deploying with gunicorn.

CMK-22721

Change-Id: Ief84abe3f6f83522be9445749e7e8666745d733a