16577 FIX Fix setup earch: Internal error: You called this page with an invalid host name
When using the setup search as non-admin user, a search could result in a crash
report with the message "Internal error: You called this page with an invalid
host name". This was caused during the computation of a single host search
result and subsequentially made the whole search fail. The error handling has
now been changed so that only the faulty search result is skipped.

CMK-22600

Change-Id: I868d51c769c5ebbeee0d10de68e0cef74544f6a0