17833 FIX Set auto focus on first form element of slidein
When opening a slide-in, the document focus was not automatically set to the form it contained.
As a result, it was not possible to use native browser features such as page up/down to scroll or tab stop to move to the next form element.

This has now been fixed and the first form element is automatically focused when a slide-in is opened.

Change-Id: I4c0b5256acd14d3a4b24ef43c85895b5d70df9c3