17389 FIX runas: Concretize ruleset
The previous wording of the _Plug-ins, local checks and MRPE for non-root users_
ruleset suggest that the main purpose is to run plugins as another user.
However, it doesn't set focus on the fact that the specified paths should contain
_additional_ agent plugins/local checks/MRPE configurations outside of the usual
agent folders.
This is now expressed more precisely with the new title _Additional plugins,
local checks, and MRPE configuration_.

The feature to run the specified plugins under another user is still present,
but it's actually optional, since you can decide to leave the field empty to
just run the plugins under root.

CMK-22627

Change-Id: I986de3d2280a84c333748f91f21e46e7be292d08