runas: Fix usage for empty user field
When leaving the user field empty in the "runas" bakery ruleset,
the agent will just ignore the directive and run plugins under the
specified directory directly.
However, in case of a non-root agent deployment, this contradicts
with the statement that this will lead to the plugins executed by root.
This is now corrected.

Change-Id: I1b075874aff11de0b00fc71ae408aef60d506e50