ci: add correct, windows only PATH
see:
https://bazel.build/reference/command-line-reference#common_options-flag--enable_platform_specific_config
https://bazel.build/run/bazelrc
"if the host OS is Linux and the build command is run, the build:linux configuration will be automatically enabled. Supported OS identifiers are linux, macos, windows, freebsd, and openbsd. Enabling this flag is equivalent to using --config=linux on Linux, --config=windows on Windows, and so on."

TESTS: manual, bazel build on Windows Node + local on LInux

Change-Id: Idafddf768dd34916f55e2a4933d619be85a45586