tests.conftest: Add --dry-run argument to simulate test execution
With --collect-only you can run a collection and get a list of tests
that would be processed by pytest. Unfortunately, that list will include
tests that are not actually executed but skipped. Using --dry-run, you
can trigger a simulated run, which will mark all tests that would be run
as XFAILED (and skip all tests that would be skipped anyway).

The XFAILED tests will have the reason set to *** DRY-RUN ***, so in case
you have any existing XFAILED tests, you will be able to distinguish them.

Change-Id: I708b8ba19304c856148053b01eff0f6655c61d6c