17896 bi: fix log level for compilation logs
Setting the log level in the global settings was not working correctly.
This should now be fixed without any user intervention. The issue with
the initial fix was that "cmk." is not need as all child loggers are
already prefixed with this.

CMK-22779

Change-Id-Ref: I5c52ca01f0beadacee4e5166ab7cb897524e8525
Change-Id: Icb41598256716b7e79f651b37126e7072a362f2d