17879 FIX Distributed piggyback: Central hub is required for distributed piggybacking to work for remote sites
This werk is relevant if you want to use the distributed piggyback mechanism.

The piggyback-hub of the central site needs to be enabled if any site in the distributed setup wants to use distributed piggybacking, irrelevant of the fact whether the central site is involved in the distribution.
Any configuration made via the web interface as introduced in [werk #17877](https://checkmk.com/werk/17877) will now be validated to that effect.

BETA-69

Change-Id: Icb5ccad9a35c3a62df03ca9af2612a8b943c69af