Get rid of json string encoding slash escaping
The slash escaping was originally added with commit 023f4bb5 to the UI
code base. While it was a counter measure for the XSS attack vector, it
was not done in the right place. Since we have improved our HTML code
rendering over time, we can now do the escaping in the right place and
get rid of this custom JSON string handling again.

Change-Id: I0e63b2192ccb00b3904c855b090cdf8e84f6a8ca