FIX 17789 mknotifyd: Duplicate services for the same sites
This werk fixes duplicate entries of *OMD Notify Connection* introduced by werk #17380.

The previous werk introduced a change that intended to fix a issue with service names and IP adresses,
which introduced a duplication bug when hostnames were used to define the connection.

This werk fixes the duplication and removes the services containing the hostnames in favour of
the services containing the site names.

SUP-23126

Change-Id: Iab7c4b2300ae4f96324e09208b6cc7125f02525f