17912 FIX Cloud Quick Setups: verify if service discovery in distributed monitoring
Prior to this werk, the Cloud Quick setups always attempted to run a service discovery during the
save process, regardless of whether the targeted site was online or not. This werk changes this
behavior by first checking the site status and skipping the service discovery if the site is unreachable.

CMK-22573

Change-Id: Ic247819c3f43128561c9fa48da9ed8454ac2e517