15362 FIX Improve autocompleter in vuejs context
Checkmk is currently in the transition to a new front end rendering. This means
some forms are still in the old world, while some (quicksetup and notification)
are already in the new world.

The new world tried to implement the hosts filter in notification slightly
differently than before. With this Werk we try to align the behaviour, but also
adapt other places.

Before this was a text box with auto suggestions, now it's a dropdown with the
possibility (depending on the context) that the user input may be presented as
an additional element to choose.

With that mechanic we can also greatly simplify the labels interface:
Previously it was required to show error messages when the user added invalid
values, now the user can no longer choose those invalid values.

Along this changes we also improved the visual appearance of drop downs in
vuejs context.

Change-Id: If1eb671e91757cec4ba6a54b00b08996269b6d2d