17915 Fix cmk/site label always reporting current site
We previously set the "cmk/site" label in a way that it always returned
the SiteId of the site currently reading the config.

For example when the agent bakery triggered an agent baking on site XYZ,
we would ask the ruleset matcher which agents are affected by which
rule. Since the ruleset matcher reported all "cmk/site" labels as
omd_site(), the bakery applies all rules for "cmk/site" == XYZ and
ignores rules with "cmk/site" set to something else.

We initially tried to fix the bug by allowing explicit labels to
overwrite builtin tags under the assumption that non-central sites
have their "cmk/site" label set in the explicit labels but this is not
the case (Werk #17335).

We revert the label matching order change and make the builtin labels
site-dependent instead.

SUP-20807
SUP-22570

Change-Id: I7185495d112ff8029cab849c2627d711de3f79ce