Adapt implementation of FormLabels
Previously the FormLabels element tried to reimplement the 2.3.0
behavior, but missed some crucial points.

Now we decided to implement it in a different way, so the once added
labels can not longer be edited, but only be removed. This makes error
reporting much easier for both the developer and the user.

Previously you could add tags nonstop:

    tag1:value1[enter]tag2:value2[enter]

hits would add the two tags. Now you have to open the dopdown first to add
another tag:

    tag1:value1[enter][enter]tag2:value2[enter]

CMK-22773
CMK-22884

Change-Id: Idcc76c54c7c9fceec1c44f1a93078b4df611aca6