Skip new RRD conversion integration test for SaaS edition
The test attempts to create rules from rulesets that are not available for SaaS
in our test setup. While this could easily be changed, the functionality under
test does not apply to SaaS. The test checks that RRD files in Nagios format can
be converted to the CMC format after upgrading from the raw to a commercial
edition. This scenario does not apply to Saas. Therefore, we will skip the test.

Change-Id: I52b05ff72daac5c8aadf206b0b5a7832116d0c5e