Integration tests: Wait until socket is reachable after restarting aut. helper
There is no guarantee that the automation helper service is reachable directly
after `omd restart automation-helper` is complete, since it takes a moment for
`gunicorn` to set up the socket. We have observed a composition test failure
that is likely caused by not waiting for the service to be reachable after
restarting the automation helper.

Note that in master, these changes are already in place. They were part of the
switch from `gunicorn` to `uvicorn`.

Change-Id: I07ddc9a6bdfbde091446b87d30b7f5aeecf2e7bf