17862 FIX GCP Quick Setup: Test connection
Similarly to the special agents for AWS and Azure, the GCP special
agent now accepts the option '--connection-test' to test the connection
with given credentials.

This new option is used in the GCP Quick Setup to fail early (on stage
one) if the connection attempt was unsuccessful and to provide users
with an according error message.

CMK-21969

Change-Id: I68c37b057286f7faa7f7f104a41618c7f0a9d5f4