Add wrapper for bootlin `ld`
Add wrapper for `ld` that removes the `--sysroot` argument.
`configure`, used with `foreign_cc` searches for the tools under
their standard names (`gcc`, `ld`, ...) and passes them all the flags
defined on the toolchain.  However, the `--sysroot` argument is not
necessary for the linker and worst, it makes linking fail.

CMK-22570

Change-Id: Ib8d0bb4e1fd8f72923a14a6ec3a0aafed5952fa1