17917 Wrap info (i) trigger with last word
In various quick setup instances, we render an (i) icon to
toggle a pop-up help field. This (i) would be wrapped onto
the next line by itself.

With this werk, we fix the wrapping behavior to always make
it stick to the previous word.

We achieve this by rendering the help by the checkbox allowing
more granular layouting.

Also indent wrapped label texts to same level as first line
after checkbox.

CMK-22719

Change-Id: Ibadab75111f61c8791a80b04a7d25e2b40d154ee