fully replace smart with 2.3.0p30 version
This agent was deployed in version 2.3.0p19 and has not changed since. We also promised no more changes to the agent in Werk 17732, so this agent should have been deployed in the first place.

The reason for this change are missing temperature services for SCSI devices. From what I recall, this is the same reason for the revert with ID Ia70db5d1b44e30036886a320458c9de7926a4990 .
This revert was not present in the 2.4.

As a precaution, I am reverting everything that happened in the 2.4, e.g., this commit simply does:

```
git checkout v2.3.0p30 -- agents/plugins/smart
```

and then replaces the version header.

CMK-22353

Change-Id: I3ba61619c15f31dd5b045c576f9940141f5b5b76