17716 check_cert: service output enhancement
The service summary and details have been enhanced
to provide more information and a better user experience.

The host to test, and the port are now included in the
service details to provide a better overview of the monitored endpoint.

Response time is no longer provided in the summary and has been
moved to the details as a fixed item, only appearing in the
summary if it triggers an alert. This provides a more concise
summary while keeping this important information available in
the details. In addition, the configured thresholds now appear
in the output when an alert is triggered.

Subject Common Name (CN) now always appears as a fixed entry in
the summary and details.

The issuer common name (CN) now always appears as a fixed item
in the details.

Previously, both the number of days remaining until certificate
expires and the expiry date were provided in the summary and details.
The certificate validity expiration date (notAfter date) is no longer
provided in the summary and details, and only the number of days
remaining until expiry is provided. This gives a more concise output
and is now consistent with the output of the httpv2 service.

Previously, if an alert was triggered,the output was only displayed
in the summary. For a more consistent experience, the information now
appears in both the summary and the details.

Change-Id: Ic7603566b84fb00e409833feff9c7439515edf55