17918 Fix element dragging not working on Firefox
We are using a new rendering framework to render forms on certain pages.
There, it was not possible to reorder list elements by dragging on
arrows next to the element.

Currently this bug is limited to the dynamic host management,
notifications and the "Quick Setup" pages.

With this Werk, we fix the underlying bug and restore the dragging
behavior on these pages on Firefox.

Change-Id: I245d7da2def2822182c07432e372bee7a39fde8f