expand comment
During the discovery, '_item_key' is added as a discovered parameter via
`Service(..., parameters={'_item_key': ...})`. During the `autochecks`
update action this key will be removed, unless the rule set contains the
'_item_key'. Moreover, the rendering of the rule set in the GUI will
only be rendered correctly, if '_item_key' is in the rule set.

Change-Id: I56835ca787f654cc94abe66248c95b62e4cf3ea8