15364 FIX Label selection: Do not suggest labels for already selected label key
When selecting labels in the new notification quick setup it was
possible to select labels although an already selected label had the
same prefix.

But it does not make sense to select label `color:red` if `color:green`
is already among the selected labels.

Now Labels with the same key as already selected labels are no longer
suggested and can not be added.

CMK-22884

Change-Id: I422e1c0f7a4516653dde8f1fda6bdb87b60cc130