17920 Stop daily deprecation tests from crashing
Checkmk registers a cronjob to regularly test the configuration for
deprecations and notifies the responsible users.

This job has the side-effect of clearing user messages which are not
valid anymore (their configured expiration date has passed).

This message cleanup expected to be triggered with an active logged in
user causing the cronjob to crash, reporting a crash report.

In this Werk we fix the underlying cause for the crash allowing the
cronjob to continue.

CMK-23107

Change-Id: I2896e666ac4d650577c85bbd2f736bbd22528d5d