FIX 17789 mknotifyd: Duplicate services for the same sites
This werk fixes duplicate entries of service *OMD Notification Spooler connection* introduced by werk #17380.

The previous werk introduced a change that intended to fix a issue with service names and IP addresses,
which introduced a duplication bug when hostnames were used to define the connection.

This werk fixes the duplication and removes the services containing the hostnames in favour of
the services containing the site names.

SUP-23126

Change-Id: I50df5f0b73bb55063b7cdbfb4c4536c2d38802b1