Introduce dedicated CI for cmk-update-agent
Essentially moving linting, unit test and so on to dedicated job.

This is as incorrect as it was before in the sense that during packaging
of the PyInstaller executable we use another set of dependencies which
might be incompatible with what we lint and test in the CI using our
standard virtual environment.

Change-Id: I1349284cbf1df9a17da60efd791de1aea41a9594