Non-root agent deployment: Rework sudo mechanism
The sudo mechanism as function `sudo_as_root` turned out to not be flexible
enough. E.g., it was incompatible with 'timeout' (or 'waitmax'),
which expects an executable and that's incompatible with a shell
function.

It's not replaced with 'SUDO_AS_ROOT', a simple environment variable that
can be placed where needed.

CMK-21494
CMK-22342

Change-Id: Iaa2c80a3c4f24cd593f42c637944c1481863db23