Pull dependency on active_config out of watolib.simple_config_file
The origin of the change is in cmk/gui/watolib/simple_config_file.py.
Everything else is mechanically following that change.

This removes another global proxy dependency from a widely used utility
class hierarchy which makes dependencies clearer and enables us to
optimize the dependencies of unit tests later on.

Change-Id: Ibbfe6088190904bc2c9223181cf33108278a4dca