17995 ESX vSphere: Opt out of template monitoring
VM templates have been added to monitoring and their state was always OK.

This change introduces a ruleset for discovery parameters to disable service discovery for VM templates.

To take advantage of this change a service rediscovery is required.

This was also asked for in the [ideas portal](https://ideas.checkmk.com/suggestions/448538/vmware-skip-vm-templates) and discussed in the [forum](https://forum.checkmk.com/t/query-vcenter-how-to-suppress-template-vms/20911/2).

CMK-20220

Closes #768

Change-Id: I1ff607929d42bf586f7e423ef482b9fa7a468d84