Consistently use aggregate initialization for sockaddr_in
The sin_zero field has been removed from the POSIX spec, but it's still
there in the Linux headers.  There is contradicting information floating
around if zeroing out that field is necessary or not.  The easiest way out
of all this chaos is to just let C++ do its thing...

A few tiny clang-tidy/IWYU fixes for the files being touched on the way.

Change-Id: I1ef36d9445503a23db4c81877b2ee58c00f87203