testlib.site: Allow ignoring running processes in Site().stop()
The current check for running processes when the site is stopped is
not very specific and will fail on any process containing the site
path, regardless if it is an externally called process or not.

Although this works fine on dockerized CI runs, it can cause trouble
for local debug runs where you have additional processes running in
the site context, that may not be stopped when you call omd stop.

Therefore this change allows skipping the check by passing the new
argument --ignore-running-procs.

Change-Id: I6578828eb751771f485b011811a8bd27a70ec10c