schemathesis tests: Fix invalid schema updates
We were doing some schema updates that are no longer applicable and
caused an exception, so those updates have been removed.

Apart from this, the function to update the schema has been made
more robust so that it will now warn when a schema is not found but
not raise an exception.

Change-Id: Ie8135b3af9b8adc4c1f36203e2c9e03363763b0b