17863 FIX Quick Setup: Prohibit reuse of locked passwords
When configuring a cloud Quick Setup, users can choose a password from
the password store. With this Werk it is now prohibited to select a
stored password which is locked by, i.e. was created with, another Quick
Setup instance.

Selecting a password of this kind results in an according error message
on stage one.

CMK-22865

Change-Id: I8631290279ff9fc23a41d893bddea503a0158a97