Set default timeout for requests
The monitoring core usually cares for the timeout for checks. (Default
is 60 seconds). In other scenarios (e.g. Service Discovery) this is not
the case and therefore this can lead to processes running forever.
I googled a bit and sometimes people configure the timeout to be 90
seconds, so 900 seconds should not cause any troubles. It is also very
unlikely that this happens in real world legit scenarios.
In my tests these long running processes also did not cause problems, so
I think 15 minutes is OK.

Change-Id: I64f31de393350defeb9019a4fe6398fc4a16c447