bi(compiler): use multiprocessing for compilation
We can compile aggregations in parallel to improve performance. We're
using `imap_unordered` as the results do not rely on each other.

See the "HACK" comment in regards to the specific implementation. The
multiprocessing support in Python has some limitations with
serialization when passing objects to a pool process.

CMK-22785

co-authored-by: Andreas Boesl <andreas.boesl@checkmk.com>

Change-Id: I6cb6cbd637fe5a39e70742c83c884e08b15f5168