Set gRPC keepalive time to 10sec to enable the gRPC keepalive timeout
... pings are sent after 10sec of no read operations on the connection
but only if there is at least one pending gRPC call.
Only if this flag is set then Bazel times out a connection if it does
not receive a ping reply after this much time.

See also https://grpc.io/docs/guides/keepalive/ and a very long
discussion with the author of the bazel remote cache being involved at
https://github.com/bazelbuild/bazel/issues/11782

CMK-23131

Change-Id: I3240c331b4920a83d69f601fdfa9b6e1eda2edf6