18005 FIX hp_proliant_da_cntlr: Convey meaning of state "other"
If an HP Proliant device returned a state "other" in its SNMP data, the service state in Checkmk was reported as "UNKNOWN".

This state in the SNMP data points to a problem with the HP instrument agent that cannot determine the state.

Now, it gets reported as "WARN" with the notice that an administrator can address the problem by upgrading the instrument agent.

SUP-23523

Change-Id: I6f01a8e6aa5dc6d9459de34e41f2def630c8c277