17898 hp_proliant_da_cntlr: handle missing data
Sometimes we receive insufficient data from the snmp agent output to
properly run the check. Instead of crashing, we are now returning a
result with an unknown state.

NOTE: we opted to check against a zero value for each of these values
because we still want to crash if the mapping upstream updates.

CMK-22110

Change-Id: I03ce6814bff6a591825dd97c6dc711e2221bdc3d