quick_setup: background job running indication
Under certain undetermined circumstances, a background job may continue
to report active status even after all subtasks have been completed,
giving the user the impression that the process is hung. A loading
indicator is added to the BackgroundJobLog component that will be
active as long as the associated background job is active.

CMK-23299

Change-Id: I0aa37566cfdf17a111bdd0839303e8c91d90cc47