Sanitize credential tuple before raising exception
I am confused with all these SNMP backends...

With Werk #11709 we introduced the inline backend or did we replace the
inline backend?
With Werk #11929 the inline Backend uses NetSNMP again (instead of PySNMP)
Werk #12703 then dropped PySNMP and now the classic backend is used?

I was not able to trigger this in a real world scenario (time-boxed), so
we do not treat this as vulnerability.

CMK-23114

Change-Id: Ifb3b10ae0a8dcb2739cb5b601d9d940f01dc5d04