Build all editions before triggering post submit heavy chain
... to make the post submit heavy chain a only testing chain and reduce
the runtime of the triggered jobs as well as the API calls requesting
the package build status. Furthermore the used executors are freed
faster as they are not just waiting for the package to be available but
only run the tests.

Follows 8e8300f496553e59b79bf654839bee3e640f1703

CMK-23344

Change-Id: Iab3d0c2b4943730dc9fb0c9f164c0748376d93c0