17952 FIX REST-API: Improve documentation for rule endpoints
This Werk improves the REST API documentation to better reflect the actual
request and response schemas.

Specifically, this means:

* marking `host_labels` and `service_labels` as deprecated in the create and
  update endpoints
* removing them entirely in the response schemas for all rule endpoints

This does not change the behaviour of these endpoints, as the shift to
`host_label_groups` and `service_label_groups` already happened in
[Werk #16045](https://checkmk.com/werk/16045).

It is recommended to migrate to label groups, as the old way of configuring
labels will be removed in a future version of Checkmk.

Change-Id: I50201e4e46aa2e41d1b551c8953b76c7e49ffeb5
JIRA-Ref: SUP-22980