18008 FIX vsphere: Fix options for displaying power states
vsphere: Fix options for displaying power states

In version 2.4.0, the default value in the configuration for the following options was changed:

* "_Display ESX Host power state on_"
* "_Display VM power state additionally on_"

However, the special agent did not accept this new value leading to a critical result in the "_Check_MK_" service complaining about an invalid choice.

This change adds the value as a valid choice.

CMK-23363

Change-Id: I3571d6459acb0a6edf90187c9e458e49219ed787