17539 Bakery accepts Windows agent installer files deployed with MKP
Previously, even if the Windows agent installer file was
correctly deployed with MKP, the Bakery would ignore it.
This issue has been resolved in the current release.

Important. It is now allowed—and encouraged—to include unsigned
MSI files in an MKP package. However, if a signed MKP must be
deployed for any reason, the author of the MKP must also provide
an unsigned patch file named `unsign-msi.patch` alongside the
installer file.
Otherwise, the customer may encounter issues during installation
on the target system.

Change-Id: I5b128321b84c0903de5f9185984bee1786a1cfd0