gui_e2e: fix activate_changes flakiness
Problem was that changes activation triggers a reload action of
"Activate Changes" page that was done before changelog url was loaded.
So, on reload, the browser navigated to the previous page instead of
reloading the "Activate Changes" page.

Waiting for the changelog url before performing activation does the
trick.

CMK-23367

Change-Id: Id5fdbecd92d7558c2ebd73ef785d5c6148a8bbda