17980 FIX Trust system wide configured CAs warning on activate changes
Before this change, when making use of *Trusted certificate authorities for SSL* with *Trust system wide configured CAs* set and having system wide certificates with permissions that would prevent Checkmk reading them. Upon activating changes, a warning would be displayed and a stack trace output to the site's *web.log* file.

As of this change, this warning is no longer displayed and a more easily parsed error is now added to the site's *web.log* file.

Change-Id: I52545f508d5e138b9ef7e8bf3a82e3e393f6a334