testlib.site: Allow copying results in uncontainerized runs
It is useful for debugging to copy the same log files
in uncontainerized runs. This way, if you drop the site
or run other tests afterwards, the log files are preserved
and can be examined later.

To do so, the RESULT_PATH environment variable must be set.

This change also turns result_dir into a property.

Change-Id: I3d0dced63118e288f4f44c8451f2f2335dd318da