18051 FIX esx_vsphere_datastore_io: Discard metrics with negative values
If an ESX host or vCenter cannot obtain data for a metric, this lack of data is indicated by the value "-1". Without handling such values they caused a ValueError in the check plugin.

For individual datastores, the check rsult is simply set to UNKNOWN.

For the datastore summary, such items are excluded from the aggregation.

CMK-22843

Change-Id: I5bcd60f082f3555d630921054f90213e3c02c272