Improve color consistency in main dashboard of commercial editions
In non-CRE sites, the main dashboard shows the host and service statistics
dashlets right next to the total host and service problems graphs. In this
change, we adjust the colors of some of the metrics shown in the problems
graphs, st. they match the colors of their counterparts in the dashlets.

CMK-23411

Change-Id: I2d8de917e0d15bbacbdbebc0d4b6545c541b5fdc