activate changes: inline some of mkp-tools logic
With this we prepare for the fact that we have
to execute these actions even if the mkp-tool
is not the one to observe the changes.

The only change in behavior here is that we no
longer optimize for the case where none of the
MKPS that were enabled or disabled contain any
files at all.

Change-Id: I530c76c76eae30c16a4801e14d48eddc2d0a4661