Merge "ci" and "local" linux_x86_64 toolchains
This is less code and less code duplication.

But more importantly, we need to maintain a fine-grained control
over the order in which the toolchains are resolved.  However,
toolchains registered using commandline arguments always have
the highest precedence.

This is unwanted for two reasons:  This prevents us from ever
reducing the precedence of the CI toolchain.  This also makes
the precedence of the CI toolchain different from the local
toolchain, which can result in bugs that are not easily
reproducible outside of the CI environment.

CMK-20981

Change-Id: I4d05f52e8de841169f7b54f99989ae0dd6c8baf4