17905 FIX Linux agent: Missing output when sudo not installed
This Werk fixes a regression that occurs since Checkmk 2.4.0 and Werk #17901.

With the introduction of the non-root agent, we replaced all `su` calls with equivalent
`sudo` calls in the Linux agent.<br>
We did this to enable a non-root agent user to be able to run these commands (when permitted).

However, since `sudo` is not available on every system, this lead to missing data since
the respective commands did fail.

We now fall back to the usage of `su` whenever `sudo` is not available.<br>

Note: We recommend to have `sudo` installed, though. We recently observed that the usage of `su`
in the Linux agent may lead to systemd user sessions starting that may have a performance impact.

CMK-23462

Change-Id: Ie1dfb9ff61947c824960a8d33b21ab08a8be5e99