17922 Stop autocompleter crash with too many items
The reworked dropdowns with autocompleted suggestions present in the
Quick Setup, Notification and DCD pages led to a crash with an
unexpected error when too many items were suggested.

This occured for example when more than 200 hosts are present and the
host filter in the new notification rule edit page is opened.

With this fix, these occurences will no longer crash.

SUP-23743

Change-Id: I493f64a44d2a11f837c6d4691848619e74ce77d1