Fix CRE composition tests
The `_host_check_template.py` is used as a template to create the so called host checks
which are executed when Nagios executs the `Check_MK` service of a host.

See `dump_precompiled_hostcheck` for a better understanding on how this
works.

The relative import in the change 7d3d21e broke the execution in the
context of the host checks, e.g.:

```
python3 /omd/sites/heute/var/check_mk/core/helper_config/3/host_checks/heute.py
```

Change-Id: I7fd2fe5ff9bb63ec8efac1e8423c0005228bce2f